fix(ci): stabilize workspace tests, resolve compiler warnings, and fix integration test panics

+++ mikrom-agent-ebpf/src/main.rs
@@ -165,7 +165,7 @@ fn try_mikrom_egress(ctx: TcContext, ifindex: u32) -> Result<i32, ()> {
             // Match protocol
             use mikrom_agent_ebpf_common::Protocol;
             let rule_proto = rule.protocol;
-            if rule_proto != Protocol::Any && rule_proto as u8 != protocol as u8 {
+            if rule_proto != Protocol::Any && rule_proto as u8 != protocol {
                 continue;
             }
